download_strategy: prioritise preferred mirrors
Some software projects, like GNU, require that mirrors be preferred over
the original URLs:

  The GNU C Library can be found on the main GNU ftp server (download
  via HTTPS,  HTTP or FTP), and its mirrors; please use a mirror if
  possible.

They save bandwidth on the original server, and often provide
significantly faster downloads.

We can maintain a list of mappings from original URLs to mirrors, and
prioritise the mirrors over the original URLs if the URL matches one of
the mappings.